# [Perf improvement] Implementing memoization for sanitizeNamespace and replaceGenericParameters funcs.