@nalexn 2023-11-14
That is a substantial performance improvement, good job! Before merging this in, I suggest an additional slight refactor of `sanitizeNamespace` in [this fasion](https://github.com/nalexn/ViewInspector/issues/268#issuecomment-1806104270).
The part where it removes `"(extension in ...)"` isn't yet confirmed to be helpful, but it'd be best to make sure your final version supports adding more patterns for prefix removal